requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiolimiter==1.1.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anthropic==0.32.0
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
assemblyai==0.26.0
astra-assistants==2.0.19
astrapy==1.4.1
asttokens==2.4.1
asyncer==0.0.5
attrs==23.2.0
Authlib==1.3.1
azure-core==1.30.2
azure-functions==1.20.0
azure-storage-blob==12.22.0
backoff==2.2.1
backports.tarfile==1.2.0
bce-python-sdk==0.9.19
bcrypt==4.0.1
beautifulsoup4==4.12.3
boto3==1.34.153
botocore==1.34.153
build==1.2.1
cachetools==5.4.0
cassandra-driver==3.29.1
cassio==0.1.8
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
clevercsv==0.8.2
click==8.1.7
cohere==5.6.2
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
coolname==2.2.0
crewai==0.36.1
crewai-tools==0.4.26
cryptography==42.0.8
dataclasses-json==0.6.7
datasets==2.19.2
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
dspy-ai==2.4.13
duckdb==1.0.0
ecdsa==0.19.0
elastic-transport==8.13.1
elasticsearch==8.14.0
email_validator==2.2.0
embedchain==0.1.120
emoji==2.12.1
environs==9.5.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fake-useragent==1.5.1
fastapi==0.111.1
fastapi-cli==0.0.5
fastavro==1.9.5
filelock==3.15.4
filetype==1.2.0
firecrawl-py==0.0.16
FLAML==2.1.2
flatbuffers==24.3.25
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.139.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-generativeai==0.7.2
google-resumable-media==2.7.1
google_search_results==2.4.2
googleapis-common-protos==1.63.2
gotrue==2.6.1
gptcache==0.1.44
greenlet==3.0.3
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.63.0
grpcio-health-checking==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.22.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ijson==3.3.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
instructor==1.3.3
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
jaraco.context==5.3.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.11
langchain-anthropic==0.1.20
langchain-astradb==0.3.3
langchain-aws==0.1.14
langchain-chroma==0.1.2
langchain-cohere==0.1.9
langchain-community==0.2.10
langchain-core==0.2.23
langchain-experimental==0.0.61
langchain-google-calendar-tools==0.0.1
langchain-google-community==1.0.7
langchain-google-genai==1.0.8
langchain-google-vertexai==1.0.7
langchain-groq==0.1.6
langchain-milvus==0.1.3
langchain-mistralai==0.1.10
langchain-mongodb==0.1.7
langchain-nvidia-ai-endpoints==0.1.6
langchain-openai==0.1.17
langchain-pinecone==0.1.2
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langdetect==1.0.9
langflow==1.0.14
langflow-base==0.0.92
langfuse==2.40.0
langsmith==0.1.96
langwatch==0.1.17
litellm==1.42.11
llama-cloud==0.0.15
llama-index==0.11.1
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-core==0.11.1
llama-index-embeddings-openai==0.2.3
llama-index-indices-managed-llama-cloud==0.3.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.0
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-parse==0.5.0
loguru==0.7.2
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.0.9
metal_sdk==2.5.1
metaphor-python==0.1.23
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
mypy-extensions==1.0.0
nanoid==2.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numexpr==2.10.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.42.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-exporter-prometheus==0.46b0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optuna==3.6.1
ordered-set==4.1.0
orjson==3.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
passlib==1.7.4
peewee==3.17.6
pgvector==0.2.5
pillow==10.4.0
pinecone-client==3.2.2
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.10.1
postgrest==0.16.9
posthog==3.5.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.8
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2-binary==2.9.9
pulsar-client==3.5.0
pure-eval==0.2.2
py==1.11.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyautogen==0.2.33
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.17.2
pylance==0.9.18
pymilvus==2.4.4
pymongo==4.8.0
pyparsing==3.1.2
pypdf==4.3.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyright==1.1.374
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.7
python-pptx==0.6.23
pytube==15.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
qdrant-client==1.10.1
qianfan==0.3.5
rapidfuzz==3.9.5
ratelimiter==1.2.0.post0
realtime==1.0.6
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3transfer==0.10.2
schema==0.7.7
scipy==1.14.0
selenium==4.23.1
semver==3.0.2
sentry-sdk==2.12.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
sqlmodel==0.0.18
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.7
StrEnum==0.4.15
striprtf==0.0.26
structlog==24.4.0
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.4.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
types-cachetools==5.4.0.20240717
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uncurl==0.0.11
unstructured==0.14.10
unstructured-client==0.25.1
upstash-vector==0.4.0
uritemplate==4.1.1
urllib3==2.2.2
uuid6==2024.1.12
uvicorn==0.30.5
validators==0.33.0
virtualenv==20.26.3
watchfiles==0.22.0
wcwidth==0.2.13
weaviate-client==4.7.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wikipedia==1.4.0
win32-setctime==1.1.0
wolframalpha==5.1.3
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
xmltodict==0.13.0
xxhash==3.4.1
yarl==1.9.4
yfinance==0.2.41
youtube-transcript-api==0.6.2
zep-python==2.0.0rc6
zipp==3.19.2
